+++ src/Providers/AuthyServiceProvider.php
@@ -24,28 +24,44 @@ use Illuminate\Support\ServiceProvider;
 class AuthyServiceProvider extends ServiceProvider
 {
     /**
+     * Indicates if loading of the provider is deferred.
+     *
+     * @var bool
+     */
+    protected $defer = true;
+
+    /**
      * {@inheritdoc}
      */
     public function register()
     {
         $httpClient = new HttpClient();
-        $mode       = config('services.authy.mode');
-        $key        = config('services.authy.keys.'.$mode);
+        $key = config('services.authy.secret');
 
-        $this->app->singleton('rinvex.authy.app', function ($app) use ($httpClient, $key, $mode) {
-            return new AuthyApp($httpClient, $key, $mode);
+        $this->app->singleton('rinvex.authy.app', function ($app) use ($httpClient, $key) {
+            return new AuthyApp($httpClient, $key);
         });
 
-        $this->app->singleton('rinvex.authy.user', function ($app) use ($httpClient, $key, $mode) {
-            return new AuthyUser($httpClient, $key, $mode);
+        $this->app->singleton('rinvex.authy.user', function ($app) use ($httpClient, $key) {
+            return new AuthyUser($httpClient, $key);
         });
 
-        $this->app->singleton('rinvex.authy.token', function ($app) use ($httpClient, $key, $mode) {
-            return new AuthyToken($httpClient, $key, $mode);
+        $this->app->singleton('rinvex.authy.token', function ($app) use ($httpClient, $key) {
+            return new AuthyToken($httpClient, $key);
         });
 
         $this->app->alias('rinvex.authy.app', AuthyApp::class);
         $this->app->alias('rinvex.authy.user', AuthyUser::class);
         $this->app->alias('rinvex.authy.token', AuthyToken::class);
+    }
+
+    /**
+     * Get the services provided by the provider.
+     *
+     * @return array
+     */
+    public function provides()
+    {
+        return ['rinvex.authy.app', 'rinvex.authy.user', 'rinvex.authy.token'];
     }
 }