+++ CHANGELOG.md
@@ -5,8 +5,20 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](CONTRIBUTING.md).
 
 
-## v1.0.0 - 2016-11-17
+## [v2.0.0] - 2016-12-20
+- Simplify code
+- Add test suites
+- Drop LTS support
+- Update Code Style
+- Defer service provider
+- Drop Authy deprecated sandbox api support
+- Push dependencies forward and require php7
+
+## [v1.0.0] - 2016-11-17
 - Commit first stable release
 
 ## v0.0.1 - 2016-11-17
 - Commit first draft
+
+[v2.0.0]: https://github.com/rinvex/laravel-authy/compare/v1.0.0...v2.0.0
+[v1.0.0]: https://github.com/rinvex/laravel-authy/compare/v0.0.1...v1.0.0